+++ docs/gist/rsf.go
@@ -4,10 +4,10 @@ import (
 	"io"
 	"log"
 
-	"github.com/qiniu/api/rsf"
+	"github.com/qiniu/api.v6/rsf"
 	"github.com/qiniu/rpc"
 
-	. "github.com/qiniu/api/conf"
+	. "github.com/qiniu/api.v6/conf"
 )
 
 func init() {