+++ rs/rs_api_test.go
@@ -4,14 +4,15 @@ import (
 	"os"
 	"testing"
 
-	. "github.com/qiniu/api/conf"
-	"github.com/qiniu/api/io"
+	. "github.com/qiniu/api.v6/conf"
+	"github.com/qiniu/api.v6/io"
 )
 
 var (
 	key        = "aa"
 	newkey1    = "bbbb"
 	newkey2    = "cccc"
+	fetchURL   = "http://www-static.u.qiniucdn.com/public/v1645/img/css-sprite.png"
 	bucketName string
 	domain     string
 	client     Client
@@ -84,7 +85,10 @@ func TestEntry(t *testing.T) {
 	if einfo.Hash != enewinfo.Hash {
 		t.Fatal("invalid entryinfo:", einfo, enewinfo)
 	}
-
+	err = client.Fetch(nil, bucketName, key, fetchURL)
+	if err != nil {
+		t.Fatal(err)
+	}
 	err = client.Move(nil, bucketName, newkey1, bucketName, newkey2)
 	if err != nil {
 		t.Fatal(err)