+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 ## CHANGE LOG
 
+### v6.0.8
+
+2015-6-24 issues [#127](https://github.com/qiniu/api/pull/127)
+
+- rs.Client 增加 Fetch 方法
+- 增加了一个用户贡献的样例（感谢 [AsGz](https://github.com/AsGz) ）
+
+### v6.0.7
+
+2015-7-11 issues [#119](https://github.com/qiniu/api/pull/119)
+
+- 修改上传域名为 upload.qiniu.com
+
 ### v6.0.6
 
 2014-6-30 issues [#114](https://github.com/qiniu/api/pull/114)
@@ -7,7 +20,6 @@
 - [#111] 修改mimetype
 - [#112] 去除chunked 上传
 - [#115] 统一user agent
-
 
 ### v6.0.5